refactor(vite): move plugin to nuxt
