+++ packages/vite/src/server.ts
@@ -8,7 +8,6 @@ import type { ViteConfig } from '@nuxt/schema'
 import type { ViteBuildContext } from './vite'
 import { createViteLogger } from './utils/logger'
 import { initViteNodeServer } from './vite-node'
-import { ssrStylesPlugin } from './plugins/ssr-styles'
 import { pureAnnotationsPlugin } from './plugins/pure-annotations'
 import { writeManifest } from './manifest'
 import { transpile } from './utils/transpile'
@@ -118,24 +117,6 @@ export async function buildServer (ctx: ViteBuildContext) {
   } satisfies vite.InlineConfig)
 
   serverConfig.customLogger = createViteLogger(serverConfig)
-
-  const chunksWithInlinedCSS = new Set<string>()
-  serverConfig.plugins!.push(ssrStylesPlugin({
-    srcDir: ctx.nuxt.options.srcDir,
-    chunksWithInlinedCSS,
-    shouldInline: ctx.nuxt.options.experimental.inlineSSRStyles
-  }))
-
-  // Remove CSS entries for files that will have inlined styles
-  ctx.nuxt.hook('build:manifest', (manifest) => {
-    for (const key in manifest) {
-      const entry = manifest[key]
-      const shouldRemoveCSS = chunksWithInlinedCSS.has(key)
-      if (shouldRemoveCSS) {
-        entry.css = []
-      }
-    }
-  })
 
   await ctx.nuxt.callHook('vite:extendConfig', serverConfig, { isClient: false, isServer: true })
